fix: fix previously task errors
