+++ src/lib/constants.ts
@@ -1,0 +1,4 @@
+export const authUrl = import.meta.env.VITE_APP_AUTH_URL;
+export const projectKey = import.meta.env.VITE_APP_PROJECT_KEY;
+export const clientId = import.meta.env.VITE_APP_CLIENT_ID;
+export const secret = import.meta.env.VITE_APP_SECRET;
