+++ src/services/routeManager.ts
@@ -1,9 +1,7 @@
-import { useLocation } from 'react-router-dom';
-import { Location } from '../interfaces';
+import { useSearchParams } from 'react-router-dom';
 
 function RouteManager() {
-  const location: Location = useLocation();
-  const searchParams = new URLSearchParams(location.search);
+  const [searchParams] = useSearchParams();
   const currentPage = searchParams.get('page');
   const currentProduct = searchParams.get('product');
   let currentLimit = searchParams.get('limit');
